+++ examples/contracts/gowdk.config.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"github.com/cssbruno/gowdk"
+	contractsaddon "github.com/cssbruno/gowdk/addons/contracts"
+)
+
+var Config = gowdk.Config{
+	Addons: []gowdk.Addon{
+		contractsaddon.Addon(),
+	},
+}